+++ docs/html/search/defines_0.js
@@ -4,5 +4,6 @@ var searchData=
   ['vma_5fbuffer_5fdevice_5faddress_1',['VMA_BUFFER_DEVICE_ADDRESS',['../vk__mem__alloc_8h.html#a7f9d5e71b70dd1a137c303a8a8262c10',1,'vk_mem_alloc.h']]],
   ['vma_5fdedicated_5fallocation_2',['VMA_DEDICATED_ALLOCATION',['../vk__mem__alloc_8h.html#af7b860e63b96d11e44ae8587ba06bbf4',1,'vk_mem_alloc.h']]],
   ['vma_5fkhr_5fmaintenance4_3',['VMA_KHR_MAINTENANCE4',['../vk__mem__alloc_8h.html#abc635c70b154cb54aa4a15cd5da55498',1,'vk_mem_alloc.h']]],
-  ['vma_5fmemory_5fbudget_4',['VMA_MEMORY_BUDGET',['../vk__mem__alloc_8h.html#a05decf1cf4ebf767beba7acca6c1ec3a',1,'vk_mem_alloc.h']]]
+  ['vma_5fkhr_5fmaintenance5_4',['VMA_KHR_MAINTENANCE5',['../vk__mem__alloc_8h.html#aae3b6bc07a36b2852bbebf5002ca8b0a',1,'vk_mem_alloc.h']]],
+  ['vma_5fmemory_5fbudget_5',['VMA_MEMORY_BUDGET',['../vk__mem__alloc_8h.html#a05decf1cf4ebf767beba7acca6c1ec3a',1,'vk_mem_alloc.h']]]
 ];