+++ src/VmaUsage.h
@@ -64,6 +64,7 @@ include all public interface declarations. Example:
 //#define VMA_MEMORY_BUDGET 0
 //#define VMA_STATS_STRING_ENABLED 0
 //#define VMA_MAPPING_HYSTERESIS_ENABLED 0
+//#define VMA_KHR_MAINTENANCE5 0
 
 //#define VMA_VULKAN_VERSION 1003000 // Vulkan 1.3
 //#define VMA_VULKAN_VERSION 1002000 // Vulkan 1.2